+++ core/template/error.py
@@ -1,0 +1,2 @@
+class QuerySelectorNotFound(Exception):
+    pass
